pyproject.toml
[tool.poetry]
name = "fullst-h"
version = "0.1.0"
description = ""
authors = ["Cavinton <enviuslxrd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
pillow = "^10.3.0"
drf-yasg = "^1.21.7"
django-filter = "^24.2"
django-cors-headers = "^4.3.1"
redis = "^5.0.6"
celery = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
